refactor: simplify options that take no values
The API was too complicated for no additional value. There is no need
to specify that an option takes no values. In fact, options should always
be able to take values. A token on the command line after an option
will always be interpreted in a first attempt as a value token. If the the
token properly parses as a value then it is used as such. If the token
does not parse as a valid value or if there is no such value topken, the
option is checked for an implicit value.